#14 added addItem component
