+++ frontend/src/components/packingItemDetails/AddItemToPackingList.tsx
@@ -1,0 +1,6 @@
+
+export default function AddItemToPackingList() {
+    return (
+        <div></div>
+    )
+}
