+++ frontend/src/components/packingItemDetails/PackingItemOverview.css
